allow empty hashref in service args
Previously, an empty hashref would be incorrectly identified as a
service to create, but since no `class` was defined, it would die trying
to load the class with Module::Runtime dying with the error `argument is
not a module name`.

So, any service that had an empty hashref in its args would die, whether
that service was configured with it, or whether overriding args with an
empty hashref.

Now, a meta hashref must contain at least one key, and all keys inside
must be meta-keys, so empty hashrefs are correctly passed-through.

Fixes #55