add better error when object config incomplete
When a service config is missing "value" or "config", it must be an
object. If it is then also missing a "class", we get a cryptic error:
"argument is not a module name". This is coming from Module::Runtime
use_module.

Now, if this situation happens, we throw an exception with a better
error message.