allow event name to be overridden when emitting
When we emit an event, we give it a name which is the same as the first
argument to `emit()` and the same as the first argument our listeners
gave to `subscribe()` or `on()`. For the most part, this is what we
want: If these three things ever differ, it means something very weird.

But for one case where we might want something different: Events
forwarded to other listeners (like a catch-all listener for all events).
In this case, we may want to know the event's original name more than we
care about the event's current name.

To enable this, we now allow the `name` field to be overridden by the
arguments to `emit()`.

Refs #12