make emitter role less polluting of consumers
Previously, when composing the Beam::Emitter role, the consumer would
get a lot of junk in their namespace, including Scalar::Util, Carp, and
Types::Standard. These things could then be called as methods on the
resulting object, which is likely not what the consuming class signed up
for.

By moving where the Moo::Role import occurs, we clean up our namespace,
which cleans up the namespaces of everyone consuming us. This could be
a problem if people are relying on this pollution, so we'll have to warn
people in the changelog

Fixes #13