custom listener class may need to be loaded prior to use; do so
subscribe() made the assumption that the listener class's code had been
loaded and invoked methods on it (isa(), new()).  For the caller, it's
more intuitive if subscribe() loaded the class's code; otherwise
if the caller doesn't load the class, subscribe will emit a

    listener object must descend from Beam::Listener

error message, which leaves the developer irate, as their code
is undoubtedly properly subclassing Beam::Listener.